+++ cookbooks/chub_plaza/environments/ssodev.rb
@@ -9,10 +9,11 @@ default_attributes(
     :host => "stats02.nexus.commercehub.com"
   },
   :chub_plaza => {
-    :root_url => "http://plaza.vagrant.dev:8080",
+    :app_context => "/secure/portal",
+    :app_root_url => "http://plaza.vagrant.dev:8080/secure/portal",
     :cas_server_url => "https://ssodev1-castle.nexus.commercehub.com",
-    :cas_service_url => "http://plaza.vagrant.dev:8080/shiro-cas",
-    :cas_failure_url => "http://plaza.vagrant.dev:8080",
+    :cas_service_url => "http://plaza.vagrant.dev:8080/secure/portal/shiro-cas",
+    :cas_failure_url => "http://plaza.vagrant.dev:8080/secure/portal",
     :census_url => "https://ssodev1-census.nexus.commercehub.com:8443",
     :vault_url => "https://ssodev1-vault.nexus.commercehub.com:8443",
     :hazelcast => {