+++ environments/ssodev1.rb
@@ -13,7 +13,7 @@ cookbook_versions({
   "chub_census" => "= 0.1.9",
   "chub_java" => "= 0.1.1",
   "chub_openldap" => "= 1.12.16",
-  "chub_plaza" => "= 0.2.1",
+  "chub_plaza" => "= 0.2.2",
   "chub_sensu_sso" => "= 0.1.17",
   "chub_vault" => "= 0.3.5",
   "cron" => "= 1.2.6",
@@ -59,9 +59,9 @@ default_attributes(
       :api_key => "3f602c6a12c6479abdfca394c9ce1ddc"
     },
     :loadbalancer_urls => {
-      :vault => "https://ssodev1-vault.nexus.commercehub.com:8443/api-docs",
-      :census => "https://ssodev1-census.nexus.commercehub.com:8443/api-docs",
-      :castle => "https://ssodev1-castle.nexus.commercehub.com/login",
+      :vault => "https://ssodev1-vault.nexus.commercehub.com:8443/ping",
+      :census => "https://ssodev1-census.nexus.commercehub.com:8443/ping",
+      :castle => "https://ssodev1-castle.nexus.commercehub.com/admin/ping",
       :plaza => "https://ssodev1-plaza.nexus.commercehub.com/buildInfo"
     }
   },